Implement remove tab
